Avoid validating content-length values for 304 responses in oghttp2.
According to https://httpwg.org/specs/rfc7230.html#rfc.section.3.3.2, a server
may send the content-length header in a 304 response, which does not have a
body. The content-length header in this case indicates the body length the
server would have sent in the corresponding 200 response. Therefore, this CL
updates oghttp2 to not check for received content-length header value mismatch
in 304 responses. This update is important for several integration test cases.

PiperOrigin-RevId: 422905007