Reject 304 responses with bodies in oghttp2.
This CL updates oghttp2 to track, based on a stream's headers, whether the
stream can receive a body following those headers. That tracking is used to
reject 304 responses with bodies in this CL and can be expanded to other
scenarios (e.g., with other response codes or request methods) in the future.

ProtocolIntegrationTest.304WithBody with oghttp2:
http://sponge2/b1fcc5f6-9628-4579-b27b-25949278af8e

PiperOrigin-RevId: 423337666