Automated g4 rollback of changelist 425943183.
*** Reason for rollback ***

Fix forward for Chromium.

*** Original change description ***

Automated g4 rollback of changelist 425681702.

*** Reason for rollback ***

This causes test failures in Chromium.  I would normally fix it, but currently QUICHE roll to Chromium is blocked on another CL and there are two other CLs that are complicated to roll.  Please allow me to roll this one back for now, I'll be happy to help with debugging after I am able to roll the latest QUICHE into Chromium.

Error is:
[ RUN      ] HeaderValidatorTest.NameHasInvalidChar
../../buildtools/third_party/lib...

***

PiperOrigin-RevId: 426203763