Ignore incoming HTTP/3 MAX_PUSH_ID frames.
Do not call Http3DebugVisitor::OnMaxPushIdFrameReceived().  This is not
implemented in the server, and is always an error in Chrome (because a server is
not allowed to send such a frame), and therefore will be visible in Chrome's
NetLog when the stream is reset.

Do not call QuicSpdySession::OnMaxPushIdFrame().  This removes the enforcement
on the values in successive MAX_PUSH_ID frames, but will allow us to remove some
complexity from QuicSpdySession when the flag is deprecated.

The enforcement on frame types (MAX_PUSH_ID is not allowed on any stream other
than the control stream, it cannot be sent by the server, it cannot be the first
frame on the control stream) is still in place.

PiperOrigin-RevId: 431034537