Move QuicheHostnameUtils implementation out of the platform.
Those are part of the platform API because they depend on net/base/url_util.h.  We already have half of that code copied into impl.cc file; copy the rest and move it out of platform impl directory.

(this does not actually move the file out of platform, that will be done in a follow-up CL)

PiperOrigin-RevId: 438377608