Handle new/pending streams on receipt of a GOAWAY in oghttp2.
This CL is a follow-up to cl/439588951, which handled the receipt of a GOAWAY
for active streams. This CL handles the receipt of a GOAWAY for new/pending
streams with stream IDs above the received GOAWAY's last_stream_id by adding
those stream IDs to a new set in OgHttp2Session and delivering OnCloseStream()
events for those streams in the next write.

This choice was motivated by the nghttp2_submit_request() (and thus
SubmitRequest()) API, which requires returning a stream ID even after receipt
of a GOAWAY. Note that simply closing the pending streams on GOAWAY would be
insufficient, as subsequent calls to SubmitRequest() could result in new or
newly pending streams. Therefore, this CL instead decides to unify handling for
new/pending streams by catching them in the common StartNewRequest() method.
Because the OnCloseStream() is delivered on the next write attempt, this CL
also updates want_write() to account for these GOAWAY-rejected streams.

This change increases oghttp2 parity with nghttp2.

PiperOrigin-RevId: 440989445