Remove VERIFY_AND_RETURN_SUCCESS.
This is a non-standard macro that calls VERIFY_SUCCESS then returns
AssertionSuccess.  The same result can be achieved by simply returning its
argument.  The only change is that VERIFY_SUCCESS logs the location of this
macro upon failure, but given that every macro that potentially generates an
AssertionFailure (like VERIFY_EQ or VERIFY_FALSE) already logs the location of
the failure itself, there is not much value in this intermediate location.

PiperOrigin-RevId: 441560319