Remove HTTP Datagram contexts
Since the IETF MASQUE WG decided to remove context IDs from the HTTP Datagram layer, we are removing them from our implementation. This CL is a no-op if contexts were not in use, which was always the case apart from some tests that are deleted by this CL. HTTP Datagrams are currently in use only in WebTransport, and in we only use WebTransport in production as a client, with contexts disabled. This CL is not flag protected because we currently do not have any use of HTTP Datagrams on our servers.

PiperOrigin-RevId: 442018079