Remove datagram context registration
This feature was removed in draft-ietf-masque-h3-datagram-06 so we're removing it from the codebase. To allow backwards compatibility with WebTransport servers that run draft-ietf-masque-h3-datagram-04 or -05, we're hand-crafting a registration capsule and sending it in that scenario. This CL is a no-op since we will send the exact same capsule. It is not flag protected because there is currently no use of HTTP Datagrams on our prod servers.

PiperOrigin-RevId: 442027763