Use more early returns in Balsa.
In most places, this reduces nesting.  Even where it does not, it generally
improves readability.

Also remove `else` wherever unnecessary because the `if` branch already ends
with a return, break, or continue statement.

In some places, invert `if` condition to do early return on error and make the
happy path less indented, in the spirit of go/gotip/episodes/1.  In some other
places, invert the condition to make the shortest path indented, or simply to
reduce nesting.

PiperOrigin-RevId: 443129286