Make BalsaBuffer constructors public.
Constructors should be private when there is a public factory method for
creating the object, to prevent accidental instantiation not by the factory
method.  However, this is not the case with BalsaBuffer.  This class is
instantiated using its constructors by BalsaHeaders and in tests, and there
making the constructor public allows the removal of a forward declaration, two
friend declaration together with their comment, and another comment stating
"this is awkward".

Also while working on BalsaBuffer constructors, replace absl::WrapUnique(new)
with std::make_unique<>.

PiperOrigin-RevId: 444855708