Do not recalculate BalsaHeaders::iterator_base::value_ unless necessary.
Frequently there are back-to-back it->first and it->second calls.  By making
`value_` an absl::optional<> type, it is possible to avoid recalculating it on
every Lookup() call.

`value_` is initialized to absl::nullopt upon construction.  It is copied by the
implicit copy assignment operator and copy constructor, and will contain the
correct views into `*header_`.  It is now reset upon iterator increment.  There
should be no other ways to mutate the iterator.

PiperOrigin-RevId: 446239817