Fix a race condition in QUIC TLS ticket processing
QUIC's TlsServerHandshaker class holds a non-owning pointer to a DecryptCallback object.  When shutting down, the handshaker makes a call through this non-owning pointer, which can race with other code which may be simultaneously deleting the callback, resulting in a use-after-free crash.

This CL solves the problem by switching to shared ownership of the callback.  This change is difficult to regression-test - the problem was uncovered by some in-progress changes to the LetoClient class which are not ready for review yet.

Protected by Trivial change, not protected.

PiperOrigin-RevId: 447795102