Add code count to support verification of quic_fix_on_stream_reset flag.
Flag count for gfe2_reloadable_flag_quic_fix_on_stream_reset is zero.  I believe
this is due to the code path never executing on our servers.  This change adds a
QUIC_CODE_COUNT in the common path (regardless of the value of the flag) to
measure fleet-wide execution, so that an informed decision can be made about
verifying this flag.

This is similar to cl/359286289

PiperOrigin-RevId: 448012287