Remove always passing DCHECK.
const_header_lines_iterator.idx_ is of type HeaderLines::size_type,
and HeaderLines is a std::vector, which can reasonably be expected to be
unsigned, therefore the DCHECK always passes.

This line causes compilation to fail in Envoy with GCC with error
"comparison of unsigned expression >= 0 is always true [-Werror=type-limits]".

PiperOrigin-RevId: 448264499