Drop support for draft-ietf-masque-h3-datagram-00
This was initially used by WebTransport but was deprecated from there before WebTransport shipped. Since then it's been used in MASQUE but that has now also moved on to later drafts. Therefore we are no longer using draft-ietf-masque-h3-datagram-00 anywhere so it can be fuly removed. HTTP Datagrams are not used in any of our production servers so we can modify this code without flag protection.

PiperOrigin-RevId: 448327545