Move some QUICHE presubmit checks from individual files to METADATA.common.
Also add "check_delta_only: false" to make sure entire files are scanned.  I
verified that all existing files comply (or added the non-compliant ones as
exclusions).

Of the new path exclusions, the ones I intend to fix immediately I marked with a
TODO with my name.  One that is pending a flag deprecation is marked with a TODO
and a bug number.  The rest I do not intend to fix, but moving them to
METADATA.common so that they apply to all new code, not only a proper subset, is
an improvement anyway.

PiperOrigin-RevId: 448330048