Support draft-ietf-masque-h3-datagram-09
This CL adds support for HTTP Datagrams draft-09. The main difference between draft-04 and draft-09 is the removal of context IDs and their registration, but those have already been fully removed from our codebase in previous CLs. The only remaining difference is the value of the h3 setting. This CL intentionally keeps WebTransport on draft-04: we want to keep our only production deployment off of the final setting codepoint in case there are last-minute changes to the specification. HTTP Datagrams are not used in any of our production servers so we can modify this code without flag protection.

PiperOrigin-RevId: 448347052