Add a poll(2)-based event loop implementation.
Currently, all of our code is hardcoded to assume EpollServer is used for the events I/O.  This is not cross-platform, since epoll is only supported on Linux.  This adds an API that abstracts the event loop away, and a default implementation that uses poll(2), which is the most cross-platform an API can get for this purpose.

PiperOrigin-RevId: 454932600