Create low-level socket library
Acts to hide platform-specific details.  Fills a similar role (and is written with a similar interface) as QuicUdpSocketApi, except a little more general, less specific to UDP and QUIC's usage of UDP (and not named specifically "UDP" to avoid silliness like using QuicUdpSocketFd for TCP).

Converted QuicUdpSocketApi to be based off this new more general library in the couple places where trivial to do so.  But I'm calling it out of scope for now to make this new library replicate all the functionality of QuicUdpSocketApi to back all of it or to convert existing usage of QuicUdpSocketApi to the new library.

Simple tests to make sure each function is at least exercised, but did not aim for exhaustive testing.

In a subsequent CL (cl/443757825), I still plan to add a higher-level abstraction on top of this to add RAII and deal with asynchronousness and such for TCP sockets.

PiperOrigin-RevId: 458070348