Automated g4 rollback of changelist 458070348.
*** Reason for rollback ***

Missed a compile break in one of the `#if !defined(__linux__)` sections.

*** Original change description ***

Create low-level socket library

Acts to hide platform-specific details.  Fills a similar role (and is written with a similar interface) as QuicUdpSocketApi, except a little more general, less specific to UDP and QUIC's usage of UDP (and not named specifically "UDP" to avoid silliness like using QuicUdpSocketFd for TCP).

Converted QuicUdpSocketApi to be based off this new more general library in the couple places where trivial to do so.  But I'm calling it out of scope for now to make this new...

***

PiperOrigin-RevId: 458083483