Roll forward of changelist 458083483 with fixes
Original had some minor bugs in the sections #if'ed out on Linux.  Fixed and confirmed those fixes by building and running the tests with the #if's commented up to run the non-Linux code on Linux.

(Note to self: When not developing in Chrome, there's a whole lot less presubmit support to make sure things still build on other platforms, so pay more attention to platform-specific code.)

*** Original change description ***

Create low-level socket library

Acts to hide platform-specific details.  Fills a similar role (and is written with a similar interface) as QuicUdpSocketApi, except a little more general, less specific to UDP and QUIC's usage of UDP (and not named specifically "UDP" to avoid silliness like using QuicUdpSocketFd...

***

PiperOrigin-RevId: 458235267