Add BalsaVisitorInterface::OnHeader().
This method is called as soon as each header is parsed.  This allows embedders
to validate headers and potentially signal errors earlier.  The motivation for
this new method is Http1ServerConnectionImplTest.HeaderInvalidCharsRejection
in Envoy, see
https://sourcegraph.com/github.com/envoyproxy/envoy@4fd31f2/-/blob/test/common/http/http1/codec_impl_test.cc?L1115-1134.
This test does not include the end of the header block, therefore
BalsaVisitorInterface::ProcessHeaders() is never called.  With this new method
BalsaParser in Envoy will be able to validate the headers and make this test
pass.
PiperOrigin-RevId: 460514650