Remove FrameSequence and MetadataSerializer.
FrameSequence is an interface only implemented by MetadataFrameSequence (in
anonymous namespace).  Remove the interface and make MetadataFrameSequence
public.

MetadataSerializer is a trivial class with no members and only one factory
method that creates a MetadataFrameSequence on the heap.  Remove that class.

Also remove the factory method.  Change call sites to instantiate
MetadataFrameSequence on the stack instead of the heap.

PiperOrigin-RevId: 460745625