Simplify signature of most HttpEncoder methods.
Return std::string instead of returing length and using a
std::unique_ptr<char[]>* outparam.  This simplifies call sites, and takes
advantage of small string optimization to avoid heap allocations for small
frames or frame headers.

Keep using QuicheBuffer for SerializeDataFrameHeader() because it allows a
performance optimization that is unfortunately not available for other frame
types.

SerializeAcceptChFrame() will be changed in a separate CL.

PiperOrigin-RevId: 461865044