Move RFC 9000 variable length integer encoding from QUIC to QUICHE
Since BHTTP uses these, and we're starting BHTTP in QUICHE, we need to move the read/write functions for varints to QUICHE. While I was in there, I improved some comments to refer to the published RFC, and removed a few ODR violations.

This CL is a no-op, it does not modify any existing functionality.

PiperOrigin-RevId: 467990159