Update QuicConnection to Use ConnectionIdGeneratorInterface.
A derived class of ConnectionIdGeneratorInterface is owned by whatever owns the connection. For most servers, this will be QuicDispatcher, which in turn receives the Generator from whatever owns it.

For Envoy-based consumers of Quiche, generate a derived class of ConnectionIdGeneratorInterface and pass a reference to it into the Constructor for QuicConnection.

The MockQuicConnection and MockGfeQuicConnection classes contain their own MockConnectionIdGenerators instead of taking the interface as a constructor argument, in attempt to reduce the footprint of this CL by several dozen files. It also makes it so that MockQuicConnectionUsers don't have to worry about generators, which is good for most users.

Protected by FLAGS_quic_reloadable_flag_quic_connection_uses_abstract_connection_id_generator.

PiperOrigin-RevId: 471570565