Convert quic_bounded_crypto_send_buffer to a QUIC_PROTOCOL_FLAG
Because we don't know what size customer certificates could be in the future, we don't anticipate enforcing this flag, and do not plan to verify it. Instead, convert to a QUIC protocol flag so that we have a mechanism to enforce the size limit if necessary in the future.

Protected by quic_bounded_crypto_send_buffer.

PiperOrigin-RevId: 471804475