Simplify higher-level socket interfaces
Where we previously had StreamClientSocket (with connect/disconnect) built on top of Socket (with receive/send), merge into a single ConnectingClientSocket.  I think this works better for two reasons:
1) Nothing in StreamClientSocket was particular to a streaming protocol.  The interface is just as applicable for stuff like "connected" UDP sockets.
2) The undirected Send/Receive previously in Socket make more sense for a connected socket whereas an unconnected socket is more likely to make use of SendTo/ReceiveFrom.  Maybe undirected Receive is useful for both, but that's an issue that can be figured out if/when we ever add unconnected socket support in these interfaces.

Naming note: My initial instinct is to call a connection-oriented client socket `ConnectedClientSocket`, but I decided to go with `ConnectingClientSocket` instead to avoid confusion with it referring to a socket that has already been connected.

PiperOrigin-RevId: 471878943