Automated g4 rollback of changelist 471573005.
*** Reason for rollback ***

After trying out a few approaches, I think it's better for the QUIC-LB connection ID generator to be just a direct addition to LoadBalancerEncoder. Thus, there is no need for mocking, and consequently, this change.

*** Original change description ***

Change LoadBalancerEncoder::Create to return unique_ptr.

More convenient in light of changes to who owns instances of LoadBalancerEncoder

***

PiperOrigin-RevId: 472454846