Drop buffered metadata for a stream in ForgetStream().
This CL is a follow-up to a comment in cl/471086207
(http://screen/7SxFkVearSiZiaP), which pointed out that buffered metadata for a
stream may still be sent even after ForgetStream() for that stream is called on
the metadata manager.

In practice, this is unlikely to happen in production because pending metadata
is flushed at once for HTTP/2 [1] and will be flushed at once for HTTP/3 [2],
without intervening ForgetStream() calls, but this CL fixes ForgetStream() to
drop buffered metadata for the stream as a solid conceptual fix and in case
future changes allow intervening ForgetStream() calls in HTTP/2 or HTTP/3.

[1] e.g., http://google3/gfe/gfe2/vtl/vtl_client_dispatcher.cc;l=294-302;rcl=464844844
[2] e.g., cl/470321865

PiperOrigin-RevId: 472729185