Generalize EventLoopTcpClientSocket between TCP and connected-UDP
Gives us the socket support needed for CONNECT-UDP.  As far as interface and implementation goes, connected UDP sockets are pretty much the same as TCP sockets other than one small parameter sent to ::socket().  But the expected behavior is obviously very different, so this CL is primarily added testing for UDP socket expectations.

PiperOrigin-RevId: 476185935