Fixup quic_name_lookup.cc includes to allow Windows
Just header-include differences as ::getaddrinfo() is otherwise the same between Windows and POSIX to the level of importance for tool/toy code.

Will be adding more use of this logic to QUIC toy server via CONNECT/CONNECT-UDP work, and I want to reduce the chances of causing issues if something in Chrome ends up pulling this in to its Win build via its partial dependency on the toy server code.

PiperOrigin-RevId: 481165980