Properly support RFC 9297
Our code has had support for draft-ietf-masque-h3-datagram-09 for a while, but now that it's been published as RFC 9297, we should have the code refer to the published names. This CL is mostly a functional no-op, with the exception of adding support for the datagram capsule type of 0x00. Since we never sent capsules, this doesn't change our sender behavior. On the receiver side, we'll still keep support for the old draft identifiers for a while.

HTTP Datagrams and capsules are not used in production on our servers, so this change is not flag-protected.

PiperOrigin-RevId: 484400359