CONNECT-IP encapsulated client support
This CL adds a new mode to masque_client that allows testing HTTP/3 over CONNECT-IP (as opposed to HTTP/3 over CONNECT-UDP). It creates UDP and IP headers on the fly.

This code isn't used in production, it solely exists for IETF interop and therefore does not have adequate test coverage or flag protection. Those will be added in the future when we migrate the code out of the masque directory.

PiperOrigin-RevId: 484744539