Allow handshaker to override QUIC transport parameters.
Due to randomized serialization of TransportParameters it is impossible for external ProofSource to guess the serialized output and compute the correct signature.
This CL allows ProofSource to pass back its version of transport parameters bitstring and uses it instead, if it's semantically equal.

No immediate prod behavior change: the default value of quic_transport_parameters is std::nullopt and parameters are left as is.

PiperOrigin-RevId: 489557924