Update fuzzer Packet/Proto Conversion to use 16B auth tags.
Virtually all packets (including all IETF QUIC packets) have 16B tags, so the fuzzer packet signature should match that.

Also, packets with very small payloads, once enabled, cannot work with short tags, so longer tags are a prerequisite.

The test was originally highly dependent on everything being a NullEncrypter, so that decryption would work even if things were out of sync. This now normalizes everything so that Initial packets are always sent with the version-appropriate encryption, which might involve the original Destination Connection ID, and all other encryption levels are sent with the TaggingEncrypter.

Some production code, when operating in fuzz mode, used the NullDecrypter for Forward Secure packets. This can't explicitly be replaced with TaggingDecrypter, since this is in the test namespace. So instead, this adds a singleton to ssl_global_data to refer to a factory in the test namespace.

PiperOrigin-RevId: 491481742