Retrieve the correct IPv6 address when sending an ICMP6 Echo Reply.
When sending an ICMP6 Echo Reply (cl/492531951), the packet was treated as if the IP6 headers were still present at the start of the packet. Instead it was accessing bytes 8-24 of the packet, which has been stripped of IP6 and ICMP6 headers leading to undefined behaviour.

This refactors out the destination IP retrieval and changed the unit test to be more rigid by comparing packet contents.
Logs from successful run: sponge2/103bb5f8-ebde-4d4f-a7fb-92abe1e721f5

PiperOrigin-RevId: 493947520