Plumb QuicStreamPriority through QuicStream and QuicSession.
No functional change still: the RFC9218 urgency has the same range as SPDY
priorities (0 to 7 inclusive), so they can be used interchangeably, and RFC9218
incremental bit is always set to false in production.

Note that this CL adds the capability to send PRIORITY_UPDATE frames with
incremental bit set to true.  This is not used in production, and in practice
the server does not send PRIORITY_UPDATE frames anyway.  However, this CL
will allow Chrome to send such frames.

This CL does not touch methods used for Google QUIC HEADERS and PRIORITY frames
on the headers stream, because they use SpdyStreamPrecedence directly.  These
calls will eventually be removed with gQUIC.

QuicWriteBlockedList changes will follow in a separate CL.

https://github.com/google/quiche/issues/25

PiperOrigin-RevId: 494791164