Add more PRIORITY_UPDATE tests.
Test that if QuicStream::SetPriority() is called with the priority that the
stream already has, then QuicSpdyStream::MaybeSendPriorityUpdateFrame() does not
send a PRIORITY_UPDATE frame.  This was prompted by automated mutation testing
commenting on cl/494755193 that removing the line
http://google3/third_party/quic/core/http/quic_spdy_stream.cc;l=595;rcl=491635277
does not cause any test failures.  Now it does.

Also add some ParsePriorityFieldValue tests to exercise code for ignoring inner
list dict values, and some tests for operator==() with out-of-bounds urgency
values.  Inspired by reviewers' comments and mutation testing findings on
cl/494755193.

Also add QuicSpdySessionTest to exercise parsing the incremental bit.  Note that
this is still unused, but at least parsed.

PiperOrigin-RevId: 495008021