Refactor WebTransport API to make it independent of QUIC.
This moves web_transport_interface.h types into their own namespace and directory, removes dependencies on QUIC types, and cleans the API in a bunch of other ways (notably, adds more documentation).

A notable functional outcome of this refactor is that we accept datagrams as string_views instead of memslices.  For HTTP/3 use cases, this will typically reduce one copy, since we always copy the contents of the datagram while appending the stream ID.

PiperOrigin-RevId: 499584983